+++ lib/fraction.dart
@@ -3,7 +3,7 @@ library fraction;
 export 'src/core.dart';
 export 'src/exceptions.dart';
 export 'src/mixed.dart';
+export 'src/extensions/fraction_num.dart';
 export 'src/extensions/fraction_string.dart';
 export 'src/extensions/mixed_string.dart';
-export 'src/extensions/fraction_num.dart';
 export 'src/extensions/mixed_num.dart';