+++ lib/src/extensions/fraction_num.dart
@@ -2,7 +2,6 @@ import 'package:fraction/fraction.dart';
 
 /// Adds [Fraction] functionalities to [num]
 extension FractionNum on num {
-
   /// Returns a fractional representation of an [int] or a [double].
   Fraction toFraction() {
     if (this is int)
@@ -10,5 +9,4 @@ extension FractionNum on num {
     else
       return Fraction.fromDouble(toDouble());
   }
-
 }