+++ lib/src/extensions/fraction_string.dart
@@ -2,7 +2,6 @@ import 'package:fraction/fraction.dart';
 
 /// Adds [Fraction] functionalities to [String]
 extension FractionString on String {
-
   /// Checks whether a string contains a valid representation of a fraction in
   /// the format 'a/b'.
   ///
@@ -22,5 +21,4 @@ extension FractionString on String {
   /// If you want to be sure that this method doesn't throw a [FractionException],
   /// use `isFraction()`.
   Fraction toFraction() => Fraction.fromString(this);
-
 }