+++ lib/src/extensions/mixed_string.dart
@@ -2,7 +2,6 @@ import 'package:fraction/fraction.dart';
 
 /// Adds [MixedFraction] functionalities to [String]
 extension MixedFractionString on String {
-
   /// Checks whether a string contains a valid representation of a mixed fraction
   /// in the format 'a b/c'.
   ///
@@ -23,5 +22,4 @@ extension MixedFractionString on String {
   /// If you want to be sure that this method doesn't throw a [MixedFractionException],
   /// use `isFraction()`.
   MixedFraction toMixedFraction() => MixedFraction.fromString(this);
-
 }