+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.2.0]
+
+ - Fixes on documentation and examples.
+ - Now when calling `toString()` on a fraction whose denominator is 1, only the numerator is printed.
+   For example, `Fraction(3, 1).toString()` returns `3` and not `3/1`,
+
+## [1.1.0]
+
+ - Added support for `operator [](int index)`.
+
+## [1.0.1]
+
+ - Fixed some health suggestions.
+ - Added a few examples in the repository.
+
 ## [1.0.0]
 
  - `Fraction` and `MixedFraction` have been released.